fix: decouple Card function methods into separate files per atomic design
